Refactor: ui.models + ui.widgets
- Rename ui.model -> ui.models
- Move custom widgets -> ui.widgets
- Expose all models and widgets directly, no need to remember implementation file
