+++ python/stack_of_tasks/ui/data_input.py
@@ -1,65 +1,11 @@
 from typing import Dict, Generic, Literal, TypeVar, Union, overload
 
-from PyQt5.QtCore import QModelIndex, pyqtSignal
-from PyQt5.QtWidgets import (
-    QComboBox,
-    QDoubleSpinBox,
-    QFormLayout,
-    QLabel,
-    QLineEdit,
-    QTreeView,
-    QWidget,
-)
+from PyQt5.QtCore import pyqtSignal
+from PyQt5.QtWidgets import QComboBox, QDoubleSpinBox, QFormLayout, QLabel, QLineEdit, QWidget
 
-from stack_of_tasks.ui.matrix_view import MatrixView
-from stack_of_tasks.ui.model import RawDataRole
+from .widgets import *
 
 PropType = TypeVar("PropType", bound=QWidget)
-
-
-class TreeComboBox(QComboBox):
-    def __init__(self, parent=None) -> None:
-        super().__init__(parent)
-
-        self._view = QTreeView()
-        self._view.setHeaderHidden(True)
-        self._view.collapsed.connect(self._view_height_changed)
-        self._view.expanded.connect(self._view_height_changed)
-
-        self.setView(self._view)
-
-    def showPopup(self):
-        super().showPopup()
-        self._view_height_changed()
-
-    def _view_height_changed(self):
-        sh = self._view.viewportSizeHint()
-        rh = self._view.sizeHintForRow(0) + 2 * self._view.autoScrollMargin()
-        sh.setHeight(max(sh.height(), rh))
-        self._view.window().resize(sh)
-
-    def findIndexOfObject(self, obj, role=RawDataRole):
-
-        for i in range(self.model().rowCount(QModelIndex())):
-            ii = self.model().index(i, 0, QModelIndex())
-
-            for j in range(self.model().rowCount(ii)):
-                ij = self.model().index(j, 0, ii)
-
-                if ij.data(role) is obj:
-                    return ij
-        return QModelIndex()
-
-    def setTreeIndex(self, index: QModelIndex):
-        if index.isValid():
-            self.setRootModelIndex(index.parent())
-            self.setModelColumn(index.column())
-            self.setCurrentIndex(index.row())
-            self.setRootModelIndex(QModelIndex())
-            self.view().setCurrentIndex(index)
-        else:
-            self.setRootModelIndex(QModelIndex())
-            self.setCurrentIndex(-1)
 
 
 class Prop(Generic[PropType]):
@@ -73,11 +19,7 @@ class Prop(Generic[PropType]):
 
 
 class DataInput(Generic[PropType], QWidget):
-    def __init__(
-        self,
-        parent=None,
-        flags=0,
-    ) -> None:
+    def __init__(self, parent=None, flags=0) -> None:
         super().__init__(parent)
 
         self._fl = QFormLayout()
@@ -86,8 +28,10 @@ class DataInput(Generic[PropType], QWidget):
 
     property_changed = pyqtSignal(str, object)
 
-    def _add_prop(self, l: str, w, row: int):
+    def _add_prop(self, l: str, w: QWidget, row: int):
         lw = QLabel(l)
+        if not w.objectName():
+            w.setObjectName(l)
         prop = Prop(lw, w)
         self._props[l] = prop
         if row is not None:
@@ -112,25 +56,14 @@ class DataInput(Generic[PropType], QWidget):
             w.setPlaceholderText(placeholder)
         return self._add_prop(label, w, row)
 
-    @overload
     def add_combo_row(
-        self, label, model=None, row: int = None, treeView: Literal[False] = False
+        self, label, model=None, row: int = None, combo=None
     ) -> Prop[QComboBox]:
-        pass
-
-    @overload
-    def add_combo_row(
-        self, label, model=None, row: int = None, treeView: Literal[True] = True
-    ) -> Prop[TreeComboBox]:
-        pass
-
-    def add_combo_row(
-        self, label, model=None, row: int = None, treeView=False
-    ) -> Union[Prop[QComboBox], Prop[TreeComboBox]]:
-        w = TreeComboBox() if treeView else QComboBox()
+        if not isinstance(combo, QComboBox):
+            combo = QComboBox()
         if model is not None:
-            w.setModel(model)
-        return self._add_prop(label, w, row)
+            combo.setModel(model)
+        return self._add_prop(label, combo, row)
 
     def add_matrix_row(self, label, row: int = None, model=None) -> Prop[MatrixView]:
         w = MatrixView(model=model)
