--- python/stack_of_tasks/ui/model/__init__.py
@@ -1,3 +1,0 @@
-from PyQt5.QtCore import Qt
-
-RawDataRole = Qt.UserRole + 10
