+++ python/stack_of_tasks/ui/models/__init__.py
@@ -1,0 +1,9 @@
+from PyQt5.QtCore import Qt
+
+RawDataRole = Qt.UserRole + 10
+
+from .enum_model import EnumModel
+from .numpy_table_model import NumpyTableModel
+from .object_model import ObjectModel
+from .ref_frames_model import RefFramesModel
+from .task_hierarchy import TaskHierarchyModel
