+++ python/stack_of_tasks/ui/models/enum_model.py
