+++ python/stack_of_tasks/ui/widgets/__init__.py
@@ -1,0 +1,3 @@
+from .button_dialog import ButtonDialog
+from .matrix_view import MatrixView
+from .tree_combobox import TreeComboBox
