+++ python/stack_of_tasks/ui/widgets/matrix_view.py
@@ -2,16 +2,7 @@ from __future__ import annotations
 
 import typing
 
-import numpy as np
-from PyQt5.QtCore import (
-    QAbstractTableModel,
-    QMargins,
-    QModelIndex,
-    QObject,
-    QRectF,
-    QSize,
-    Qt,
-)
+from PyQt5.QtCore import QMargins, QModelIndex, QObject, QRectF, QSize, Qt
 from PyQt5.QtGui import QPainter, QTextOption
 from PyQt5.QtWidgets import (
     QDoubleSpinBox,
@@ -22,8 +13,6 @@ from PyQt5.QtWidgets import (
     QWidget,
 )
 
-FloatDataRole = Qt.UserRole + 1
-
 if not hasattr(QSize, "grownBy"):
 
     def grownBy(self: QSize, margins: QMargins):
@@ -33,47 +22,6 @@ if not hasattr(QSize, "grownBy"):
         )
 
     setattr(QSize, "grownBy", grownBy)
-
-
-class NumpyTableModel(QAbstractTableModel):
-    def __init__(self, parent=None) -> None:
-        super().__init__(parent)
-
-        self._matrix: np.ndarray = np.identity(4)
-
-    def rowCount(self, parent: QModelIndex = None) -> int:
-        return self._matrix.shape[0]
-
-    def columnCount(self, parent: QModelIndex = None) -> int:
-        return self._matrix.shape[1]
-
-    def flags(self, index: QModelIndex) -> Qt.ItemFlags:
-        return super().flags(index) | Qt.ItemIsEditable
-
-    def data(self, index: QModelIndex, role: int = ...) -> typing.Any:
-        if index.isValid():
-            if role == Qt.DisplayRole:
-                return str(self._matrix[index.row(), index.column()])
-
-            elif role == FloatDataRole:
-                return self._matrix[index.row(), index.column()]
-
-    def valuesChanged(self):
-        self.modelAboutToBeReset.emit()
-        self.dataChanged.emit(
-            self.index(0, 0), self.index(self.rowCount(), self.columnCount())
-        )
-        self.modelReset.emit()
-
-    def setData(self, index: QModelIndex, value: typing.Any, role: int = ...) -> bool:
-        self._matrix[index.row(), index.column()] = value
-        self.dataChanged.emit(index, index)
-        return True
-
-    def setMatrix(self, matrix):
-        self.modelAboutToBeReset.emit()
-        self._matrix = matrix
-        self.modelReset.emit()
 
 
 class MatrixItemDelegate(QStyledItemDelegate):
@@ -89,7 +37,7 @@ class MatrixItemDelegate(QStyledItemDelegate):
 
     def _format_index(self, index: QModelIndex):
         s = ""
-        if index.isValid() and (f := index.data(FloatDataRole)) is not None:
+        if index.isValid() and (f := index.data(Qt.EditRole)) is not None:
             s = f"{f:.2f}"
         return s
 
@@ -102,7 +50,7 @@ class MatrixItemDelegate(QStyledItemDelegate):
         return self._editor_widget
 
     def setEditorData(self, editor: QWidget, index: QModelIndex) -> None:
-        f = index.data(FloatDataRole)
+        f = index.data(Qt.EditRole)
         self._editor_widget.setValue(f)
 
     def updateEditorGeometry(
